+++ packages/gallery/src/settings/index.js
@@ -29,6 +29,7 @@ import textImageSpace from './options/textImageSpace';
 import textBoxBorderColor from './options/textBoxBorderColor';
 import textBoxBorderRadius from './options/textBoxBorderRadius';
 import textBoxBorderWidth from './options/textBoxBorderWidth';
+import textBoxWidth from './options/textBoxWidth';
 import groupSize from './options/groupSize';
 
 export default {
@@ -63,5 +64,6 @@ export default {
     textBoxBorderColor,
     textBoxBorderRadius,
     textBoxBorderWidth,
+    textBoxWidth,
     groupSize,
 }